Apply requested changes about Trait in Store handlers
