+++ src/Adapter/Store/CommandHandler/EditStoreHandler.php
@@ -8,16 +8,17 @@ declare(strict_types=1);
 
 namespace PrestaShop\PrestaShop\Adapter\Store\CommandHandler;
 
-use Country;
+use PrestaShop\PrestaShop\Adapter\Store\Trait\StoreHandlerTrait;
 use PrestaShop\PrestaShop\Core\CommandBus\Attributes\AsCommandHandler;
 use PrestaShop\PrestaShop\Core\Domain\Store\Command\EditStoreCommand;
 use PrestaShop\PrestaShop\Core\Domain\Store\CommandHandler\EditStoreHandlerInterface;
-use PrestaShop\PrestaShop\Core\Domain\Store\Exception\StoreConstraintException;
 use PrestaShop\PrestaShop\Core\Domain\Store\Repository\StoreRepository;
 
 #[AsCommandHandler]
 final class EditStoreHandler implements EditStoreHandlerInterface
 {
+    use StoreHandlerTrait;
+
     public function __construct(
         private readonly StoreRepository $storeRepository,
     ) {
@@ -82,44 +83,5 @@ final class EditStoreHandler implements EditStoreHandlerInterface
         if (null !== $command->getShopAssociation()) {
             $this->storeRepository->updateShopAssociation($store, $command->getShopAssociation());
         }
-    }
-
-    private function assertStateCountryConsistency(int $countryId, ?int $stateId): void
-    {
-        $country = new Country($countryId);
-
-        if ($country->contains_states && !$stateId) {
-            throw new StoreConstraintException(
-                'A state is required for the selected country.',
-                StoreConstraintException::INVALID_STATE
-            );
-        }
-
-        if (!$country->contains_states && $stateId) {
-            throw new StoreConstraintException(
-                'The selected country does not contain states.',
-                StoreConstraintException::STATE_COUNTRY_MISMATCH
-            );
-        }
-    }
-
-    /**
-     * @param array<int, array<int, string>> $localizedHours
-     *
-     * @return array<int, string>
-     */
-    private function encodeHours(array $localizedHours): array
-    {
-        $result = [];
-        foreach ($localizedHours as $langId => $days) {
-            $encoded = [];
-            foreach ($days as $day) {
-                $parts = array_map('trim', explode('|', (string) $day, 2));
-                $encoded[] = isset($parts[1]) ? [$parts[0], $parts[1]] : [$parts[0]];
-            }
-            $result[(int) $langId] = json_encode($encoded);
-        }
-
-        return $result;
     }
 }
